All required fields present now
the knife object can now be completely defined from an html page. Manual data entry is no longer necessary.

test_run() contains example usage of how to parse based on a url (pretty much the same as run() ).

Checking the db for an existing knife entry does not work currently.

A test with ddg and a regular string query such as "Paramilitary 2" should be added in