"cat" or category pages now parsed
the bhq_query method now can handle a page if it is a "cat" or a
category containing multiple knives. There are usually several knives
for this page so it might be nice to allow the user to select what
they meant (maybe the first 5 knives are shown and they user can
select which one they want info on).